resume-analyser-backend/requirements.txt
flask
flask-cors
PyMuPDF
python-docx
scikit-learn
pandas
numpy
spacy
joblib
python-dotenv
pdfplumber
